field: Separate display(replay) field into a new file
This patch adds new files 'utils/field.[ch]' to put field display
related features.  It is another preparation for graph field display.
